LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-upload-expected.txt


PASS Fetch with PUT with body 
PASS Fetch with POST with text body 
PASS Fetch with POST with URLSearchParams body 
PASS Fetch with POST with Blob body 
PASS Fetch with POST with ArrayBuffer body 
PASS Fetch with POST with Uint8Array body 
PASS Fetch with POST with Int8Array body 
PASS Fetch with POST with Float32Array body 
PASS Fetch with POST with Float64Array body 
PASS Fetch with POST with DataView body 
PASS Fetch with POST with Blob body with mime type 
FAIL Fetch with POST with ReadableStream assert_equals: expected "Test" but got ""
FAIL Fetch with POST with ReadableStream containing String undefined is not an object (evaluating 'promise.then')
FAIL Fetch with POST with ReadableStream containing null undefined is not an object (evaluating 'promise.then')
FAIL Fetch with POST with ReadableStream containing number undefined is not an object (evaluating 'promise.then')
FAIL Fetch with POST with ReadableStream containing ArrayBuffer undefined is not an object (evaluating 'promise.then')
FAIL Fetch with POST with ReadableStream containing Blob undefined is not an object (evaluating 'promise.then')
